Add files via upload
# A complete mini machine learning project

This project aims to apply all the standard methodologies of data science to a real-world business problem.

The business need tackled in this project is to help stakeholders of a construction and cement factory ascertain the 
quality and strength of their concrete mix for buildings and other use cases.

An attempt was made to develop a machine learning application that can predict the comprehensive strength of concrete as 
function of the cement, the age(in days) and other complex features used in preparing it.

Comprehensive Strength explains the capacity of the concrete to withstand loads before failure.
Comprehensive strength of cement concrete is an important property of the mixture that helps in its durability and 
support for use case, as the most important strength test.

It is a complex formula of relationship between age (number of days) and other factors, hence it gives an idea about the
characteristics of the concrete.
The importance of comprehensive strength is that it ensures the overall quality of the finished product. 

The propose methodology is first to defined the business need, and suggest the analytical appraoch to use in solving the 
problem. Further, we'll proceed to identify the source of data and the data requirement for this specific business need.
Again, after we understand the data, we proceed to develop our model and evaluate it. Sequel to this, if our model meets our
business need, we deploy it after every development in the production environment.

Many considerations, such as practical considerations, ethical and security consideration of the model were considered.

The advantage of this business solution is that it will enable stakeholders to minimize loss of resources. This is beacause,
prior to production they will have known what formula of their concrete mix works and which does not work, as compared to
their benchmark even before they began their production.
This approach also aim to maximize profit as the revenue will continue to grow as losses are minimized.

This project is self-inspired and an attempt to understand the Keras library holistically.

by Blibo Albert @bliboalbert <<-github/LinkedIn